+++ packages/nns/src/errors/governance.errors.ts
@@ -1,12 +1,11 @@
 import type { GovernanceError as GovernanceErrorDetail } from "../../candid/governance";
-import type { ICP } from "../icp";
 
 export abstract class StakeNeuronError extends Error {}
 
 export class CouldNotClaimNeuronError extends StakeNeuronError {}
 
 export class InsufficientAmountError extends StakeNeuronError {
-  constructor(public readonly minimumAmount: ICP) {
+  constructor(public readonly minimumAmount: bigint) {
     super();
   }
 }