+++ packages/nns/src/errors/ledger.errors.ts
@@ -1,5 +1,5 @@
 import type { TransferError as RawTransferError } from "../../candid/ledger";
-import { ICP } from "../icp";
+import { convertStringToE8s } from "../token";
 import type { BlockHeight } from "../types/common";
 
 export class TransferError extends Error {}
@@ -7,7 +7,7 @@ export class TransferError extends Error {}
 export class InvalidSenderError extends TransferError {}
 
 export class InsufficientFundsError extends TransferError {
-  constructor(public readonly balance: ICP) {
+  constructor(public readonly balance: bigint) {
     super();
   }
 }
@@ -39,8 +39,9 @@ export const mapTransferError = (
     return new TxDuplicateError(rawTransferError.TxDuplicate.duplicate_of);
   }
   if ("InsufficientFunds" in rawTransferError) {
-    const icp = ICP.fromE8s(rawTransferError.InsufficientFunds.balance.e8s);
-    return new InsufficientFundsError(icp);
+    return new InsufficientFundsError(
+      rawTransferError.InsufficientFunds.balance.e8s
+    );
   }
   if ("TxCreatedInFuture" in rawTransferError) {
     return new TxCreatedInFutureError();
@@ -83,8 +84,8 @@ export const mapTransferProtoError = (responseBytes: Error): TransferError => {
         /debit account.*, current balance: (\d*(\.\d*)?)/
       );
       if (m && m.length > 1) {
-        const balance = ICP.fromString(m[1]);
-        if (balance instanceof ICP) {
+        const balance = convertStringToE8s(m[1]);
+        if (typeof balance === "bigint") {
           return new InsufficientFundsError(balance);
         }
       }