+++ packages/nns/src/types/ledger_converters.ts
@@ -1,10 +1,9 @@
 import type { AccountIdentifier } from "../account_identifier";
-import type { ICP } from "../icp";
 import type { E8s } from "./common";
 
 export type TransferRequest = {
   to: AccountIdentifier;
-  amount: ICP;
+  amount: bigint;
   memo?: bigint;
   fee?: E8s;
   fromSubAccount?: number[];