+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 
 - new library `@dfinity/utils`
 - more Sns related features: `notifyParticipation`, `getUserCommitment`, some Sns neurons related features and governance `metadata`
+- new generic interface `Token` and class `TokenAmount`
+- class `ICP` set as deprecated
 
 ### Build